Refactor out ChainedDreamFallingBlock
